python/requirements.txt
Flask==0.10.1
Flask-HTTPAuth==2.2.0
Flask-Login==0.2.10
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.19
SQLAlchemy==0.9.3
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.4
decorator==3.4.0
flask-heroku==0.1.8
gunicorn==18.0
itsdangerous==0.23
numpy==1.8.0
pbr==0.6
psycopg2==2.5.2
python-dateutil==2.2
pytz==2013.9
redis==2.9.1
requests==2.2.1
rq==0.3.13
six==1.5.2
sqlalchemy-migrate==0.9
times==0.6.2
wsgiref==0.1.2
